+++ forui/test/src/widgets/otp_field/otp_field_golden_test.dart
@@ -1,3 +1,6 @@
+@Tags(['golden'])
+library;
+
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';