+++ forui/test/src/widgets/select/select_item_golden_test.dart
@@ -1,3 +1,6 @@
+@Tags(['golden'])
+library;
+
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';