+++ forui/test/src/widgets/slider/slider_golden_test.dart
@@ -1,3 +1,6 @@
+@Tags(['golden'])
+library;
+
 import 'package:flutter/material.dart' hide Thumb;
 
 import 'package:flutter_test/flutter_test.dart';