+++ forui/tool/cli_generator/snippets/generate_material_theme.dart
@@ -30,7 +30,7 @@ class ToApproximateMaterialThemeFinder extends RecursiveAstVisitor<void> {
 
   @override
   void visitClassDeclaration(ClassDeclaration declaration) {
-    final name = declaration.name.lexeme;
+    final name = declaration.namePart.typeName.lexeme;
     if (name == 'FThemeData') {
       super.visitClassDeclaration(declaration);
     }